requirements.txt
streamlit
pillow
python-dotenv
jamaibase
requests
